Resolved comments
