--- frontend/kobweb-compose/src/jsMain/kotlin/com/varabyte/kobweb/compose/ui/modifiers/WordSpacingModifiers.kt
@@ -1,14 +1,0 @@
-package com.varabyte.kobweb.compose.ui.modifiers
-
-import com.varabyte.kobweb.compose.css.*
-import com.varabyte.kobweb.compose.ui.Modifier
-import com.varabyte.kobweb.compose.ui.styleModifier
-
-fun Modifier.wordSpacing(wordSpacing: WordSpacing) = styleModifier {
-    wordSpacing(wordSpacing)
-}
-
-fun Modifier.wordSpacing(length: CSSLengthNumericValue) = styleModifier {
-    wordSpacing(WordSpacing.of(length))
-}
-
